+++ src/main/java/com/game/player/command/StartGameCommand.java
@@ -2,14 +2,14 @@ package com.game.player.command;
 
 public class StartGameCommand extends BaseCommand<String> {
 
-    private final double initialValue;
+    private final long initialValue;
 
-    public StartGameCommand(String id, double initialValue) {
+    public StartGameCommand(String id, long initialValue) {
         super(id);
         this.initialValue = initialValue;
     }
 
-    public double getInitialValue() {
+    public long getInitialValue() {
         return initialValue;
     }
 }
