+++ src/main/java/com/game/player/event/InitializeScoreEvent.java
@@ -2,14 +2,14 @@ package com.game.player.event;
 
 public class InitializeScoreEvent extends BaseEvent<String> {
 
-    private final double initialValue;
+    private final long initialValue;
 
-    public InitializeScoreEvent(String id, double initialValue) {
+    public InitializeScoreEvent(String id, long initialValue) {
         super(id);
         this.initialValue = initialValue;
     }
 
-    public double getInitialValue() {
+    public long getInitialValue() {
         return initialValue;
     }
 }
